requirements.txt
numpy==1.21.2
PyQt5~=5.15
pyserial~=3.5
pyqtgraph==0.12.2
openpyxl~=3.0.7